Release/candidate2.13

See merge request !1798

+++ kadena2.0/Kadena2.0.BusinessLogic/Contracts/IUserBudgetService.cs
@@ -8,6 +8,6 @@ namespace Kadena.BusinessLogic.Contracts
 {
     public interface IUserBudgetService
     {
-        string UpdateUserBudgetAllocation(int itemID, double userBudget);
+        string UpdateUserBudgetAllocation(int itemID, decimal userBudget);
     }
 }
